requirements.txt
Flask-GraphQL
Flask-SQLAlchemy
SQLAlchemy
graphene-sqlalchemy
flask-cors